bot/pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = []
dependencies = [
    "python-telegram-bot>=20.1",
]
requires-python = ">=3.11"
license = { text = "MIT" }

[tool.pdm.dev-dependencies]
dev = [
    "typer>=0.7.0",
]

[tool.pdm.scripts]
_.env_file = "../.env"
start = "python src"
g = { call = "cli.generate:main" }
